% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A matrix or data frame representing the graph, where each row is an edge.
The first column should be the starting node, the second column the ending node,
and the third column the weight}

\item{init_node}{An integer representing the starting node for the algorithm}
}
\value{
A list containing two elements:
\describe{
\item{distances}{A numeric vector representing the shortest distances from the start node to all other nodes.}
\item{previous}{An integer vector where each element represents the previous node in the shortest path.}
}
}
\description{
This function calculates the shortest paths
from a starting node to all other nodes
in a graph using Dijkstra's algorithm.
}
