% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{An integer, one of the two numbers for which the GCD is to be computed}

\item{b}{An integer, the other number for which the GCD is to be computed}
}
\value{
An integer representing the greatest common divisor of \code{a} and \code{b}
}
\description{
This function computes the greatest common divisor (GCD) of two integers
using the Euclidean algorithm
}
\examples{
euclidean(123612, 13892347912)

}
