% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\name{Mfpkg}
\alias{Mfpkg}
\alias{Mfpkg-package}
\title{Mfpkg: A Package for Dijkstra and Euclidean Algorithms}
\description{
This is an R package that provides implementations of Dijkstra's algorithm
for shortest paths and the Euclidean algorithm for calculating the greatest common divisor.
}
\details{
This package contains two functions: \code{Dijkstra} and \code{Euclidean}.
}
