% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki Graph Dataset}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{v1}{The starting node (integer).}
\item{v2}{The ending node (integer).}
\item{w}{The weight of the edge between nodes (numeric).}
}
}
\source{
Generated for graph algorithm testing.
}
\usage{
wiki_graph
}
\description{
A dataset representing a weighted, directed graph. This dataset can be used
to test graph algorithms like Dijkstra's algorithm.
}
\references{
Dijkstra's algorithm, see \url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}.

Graph theory, see \url{https://en.wikipedia.org/wiki/Graph_(mathematics)}.
}
\keyword{datasets}
